rename and move both AltinnAppHeaders to prevent confusion
