+++ src/features/instantiate/containers/InstantiationContainer.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import { AltinnAppHeader } from 'src/components/altinnAppHeader';
 import { ReadyForPrint } from 'src/components/ReadyForPrint';
 import { DataLoadingProvider } from 'src/core/contexts/dataLoadingContext';
 import { TaskStoreProvider } from 'src/core/contexts/taskStoreContext';
 import { RenderStart } from 'src/core/ui/RenderStart';
 import { Footer } from 'src/features/footer/Footer';
 import classes from 'src/features/instantiate/containers/InstantiationContainer.module.css';
+import { InstantiateHeader } from 'src/features/instantiate/instantiateHeader/InstantiateHeader';
 import { useProfile } from 'src/features/profile/ProfileProvider';
 import { AltinnPalette } from 'src/theme/altinnAppTheme';
 import { changeBodyBackground } from 'src/utils/bodyStyling';
@@ -24,7 +24,7 @@ export function InstantiationContainer({ children }: IInstantiateContainerProps)
       <DataLoadingProvider>
         <RenderStart>
           <div className={classes.container}>
-            <AltinnAppHeader profile={profile} />
+            <InstantiateHeader profile={profile} />
             <main id='main-content'>{children}</main>
             <Footer />
             <ReadyForPrint type='load' />
